+++ src/routes.ts
@@ -3,6 +3,7 @@ import Router from 'koa-router'
 
 import HealthCheckController from './controllers/healthcheck.controller'
 import IndexController from './controllers/index.controller'
+import PrivateController from './controllers/private.controller'
 import { privateRoute, publicRoute } from './utils/route.utils'
 
 const router = new Router()
@@ -11,7 +12,7 @@ router.get('/', publicRoute('base', '/', IndexController.getIndex) as any)
 
 router.get(
   '/private',
-  privateRoute('private', '/', IndexController.getIndex) as any
+  privateRoute('private', '/', PrivateController.getPrivate) as any
 )
 
 router.get('/healthcheck', HealthCheckController.getHealthCheck)
